* added maps details, locations and distance + changed icons

+++ sources/mechta-ios-app/MechtaApp/Data/Models/CityModel.h
@@ -14,9 +14,9 @@
 
 @property NSString* name;
 
-@property float latitude;
+@property double latitude;
 
-@property float longitude;
+@property double longitude;
 
 @property NSMutableArray *shops;
